serialize: Make object mapper factories their own TypeSerializers
