+++ core/src/main/java/org/spongepowered/configurate/objectmapping/ConfigSerializable.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.spongepowered.configurate.serialize;
+package org.spongepowered.configurate.objectmapping;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
