+++ core/src/test/java/org/spongepowered/configurate/objectmapping/ObjectMapperTest.java
@@ -30,7 +30,6 @@ import org.spongepowered.configurate.CommentedConfigurationNode;
 import org.spongepowered.configurate.ConfigurationOptions;
 import org.spongepowered.configurate.objectmapping.meta.Comment;
 import org.spongepowered.configurate.objectmapping.meta.Setting;
-import org.spongepowered.configurate.serialize.ConfigSerializable;
 
 import java.net.MalformedURLException;
 import java.net.URL;
