+++ examples/src/main/java/org/spongepowered/configurate/examples/ValueReferences.java
@@ -20,11 +20,11 @@ import io.leangen.geantyref.TypeToken;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.spongepowered.configurate.CommentedConfigurationNode;
 import org.spongepowered.configurate.hocon.HoconConfigurationLoader;
+import org.spongepowered.configurate.objectmapping.ConfigSerializable;
 import org.spongepowered.configurate.objectmapping.ObjectMappingException;
 import org.spongepowered.configurate.reference.ConfigurationReference;
 import org.spongepowered.configurate.reference.ValueReference;
 import org.spongepowered.configurate.reference.WatchServiceListener;
-import org.spongepowered.configurate.serialize.ConfigSerializable;
 import org.spongepowered.configurate.transformation.NodePath;
 
 import java.io.IOException;
