+++ format/hocon/src/test/java/org/spongepowered/configurate/hocon/HoconConfigurationLoaderTest.java
@@ -28,10 +28,10 @@ import org.junit.jupiter.api.io.TempDir;
 import org.spongepowered.configurate.CommentedConfigurationNode;
 import org.spongepowered.configurate.ConfigurationOptions;
 import org.spongepowered.configurate.loader.AtomicFiles;
+import org.spongepowered.configurate.objectmapping.ConfigSerializable;
 import org.spongepowered.configurate.objectmapping.ObjectMapper;
 import org.spongepowered.configurate.objectmapping.ObjectMappingException;
 import org.spongepowered.configurate.objectmapping.meta.Setting;
-import org.spongepowered.configurate.serialize.ConfigSerializable;
 
 import java.io.BufferedReader;
 import java.io.IOException;
